+++ swift/BUILD
@@ -2,5 +2,5 @@ load("@build_bazel_rules_swift//swift:swift.bzl", "swift_test")
 
 swift_test(
     name = "tests",
-    srcs = [ "main.swift"],
+    srcs = [ "main.swift", "example.swift"],
 )
